+++ src/com/jec/commons/files/FileStats.ts
@@ -1,0 +1,87 @@
+//  DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
+//
+//   Copyright 2016-2017 Pascal ECHEMANN.
+//
+//   Licensed under the Apache License, Version 2.0 (the "License");
+//   you may not use this file except in compliance with the License.
+//   You may obtain a copy of the License at
+//
+//       http://www.apache.org/licenses/LICENSE-2.0
+//
+//   Unless required by applicable law or agreed to in writing, software
+//   distributed under the License is distributed on an "AS IS" BASIS,
+//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//   See the License for the specific language governing permissions and
+//   limitations under the License.
+
+/**
+ * The <code>FileStats<code> interface defines the default set of APIs 
+ * that must be implemented for creting objects that provides system information
+ * about files or directories.
+ */
+export interface FileStats {
+
+  /**
+   * Returns a boolean that indicates whether the associated item is a file
+   * (<code>true</code>), or not (<code>false</code>).
+   * 
+   * @return {boolean} <code>true</code> whether the associated item is a file;
+   *                   <code>false</code> otherwise.
+   */
+  isFile():boolean;
+
+  /**
+   * Returns a boolean that indicates whether the associated item is a directory
+   * (<code>true</code>), or not (<code>false</code>).
+   * 
+   * @return {boolean} <code>true</code> whether the associated item is a
+   *                   directory; <code>false</code> otherwise.
+   */
+  isDirectory():boolean;
+  
+  /**
+   * Returns a boolean that indicates whether the associated item is a block
+   * device (<code>true</code>), or not (<code>false</code>).
+   * 
+   * @return {boolean} <code>true</code> whether the associated item is a block
+   *                   device; <code>false</code> otherwise.
+   */
+  isBlockDevice():boolean;
+  
+  /**
+   * Returns a boolean that indicates whether the associated item is a character
+   * device (<code>true</code>), or not (<code>false</code>).
+   * 
+   * @return {boolean} <code>true</code> whether the associated item is a
+   *                   character device; <code>false</code> otherwise.
+   */
+  isCharacterDevice():boolean;
+  
+  /**
+   * Returns a boolean that indicates whether the associated item is a symbolic
+   * link (<code>true</code>), or not (<code>false</code>).
+   * 
+   * @return {boolean} <code>true</code> whether the associated item is a
+   *                   symbolic link; <code>false</code> otherwise.
+   */
+  isSymbolicLink():boolean;
+  
+  /**
+   * Returns a boolean that indicates whether the associated item is a fifo file
+   * (<code>true</code>), or not (<code>false</code>).
+   * 
+   * @return {boolean} <code>true</code> whether the associated item is a fifo
+   *                   file; <code>false</code> otherwise.
+   */
+  isFIFO():boolean;
+  
+  /**
+   * Returns a boolean that indicates whether the associated item is a socket
+   * (<code>true</code>), or not (<code>false</code>).
+   * 
+   * @return {boolean} <code>true</code> whether the associated item is a 
+   *                   socket; <code>false</code> otherwise.
+   */
+  isSocket():boolean;
+  
+}