+++ utils/test-utils/file-properties-utils/SampleDecorator.ts
@@ -1,5 +1,0 @@
-export function SampleDecorator():Function {
-  return function(target:any):Function {
-    return target;
-  }
-}