+++ utils/test-utils/file-properties-utils/SampleFile.ts
@@ -1,3 +1,0 @@
-import {SampleDecorator} from "./SampleDecorator";
-@SampleDecorator()
-export class SampleFile {}